In progress calendar layout and date selectors
