+++ front-end/src/Page.module.css
@@ -1,0 +1,5 @@
+.page{
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
