+++ front-end/src/components/Hero/Calendar/Calendar.module.css
@@ -1,9 +1,8 @@
-.board{
+.calendar{
   background-color: #fdfaf6;
   display: grid;
   grid-template-columns: repeat(7, 1fr);
+  grid-template-rows: repeat(5, 1fr);
   width: 100%;
-}
-.calendar{
-  width: 100%;
+  height: 100%;
 }
