+++ front-end/src/components/Hero/DateSelector/index.js
@@ -1,9 +1,11 @@
 import MonthSelector from './MonthSelector';
 import YearSelector from './YearSelector';
 
+import styles from './Selectors.module.css';
+
 const DateSelector = ({ currentDate, setCurrentDate }) => {
   return (
-    <>
+    <div className={styles.layout}>
       <YearSelector
         value={currentDate.year}
         onChange={(year) => setCurrentDate((prev) => ({ ...prev, year }))}
@@ -12,7 +14,7 @@ const DateSelector = ({ currentDate, setCurrentDate }) => {
         value={currentDate.month}
         onChange={(month) => setCurrentDate((prev) => ({ ...prev, month }))}
       />
-    </>
+    </div>
   );
 };
