+++ front-end/src/components/Hero/index.js
@@ -1,10 +1,12 @@
-import { useState } from 'react';
+import { useState, useContext, useEffect } from 'react';
 import Modal from 'react-modal';
 
 import Scheduler from '../Scheduler';
-import Calendar from '../Calendar';
 import EventPanel from '../EventPanel';
-import { DateProvider } from '../../contexts/date';
+import { DateContext } from '../../contexts/date';
+import DateSelector from './DateSelector';
+import Calendar from './Calendar';
+import { get } from '../../api';
 
 import styles from './Hero.module.css';
 import button from '../StyleComponents/Buttons.module.css';
@@ -12,6 +14,14 @@ import button from '../StyleComponents/Buttons.module.css';
 Modal.setAppElement('#root');
 
 const Hero = () => {
+  const {
+    currentDate,
+    setCurrentDate,
+    setDayFrames,
+    updateCalendar,
+    setUpdate,
+  } = useContext(DateContext);
+
   const [dayData, setDayData] = useState({ type: 'empty' });
   const [modalIsOpen, setIsOpen] = useState(false);
 
@@ -21,6 +31,21 @@ const Hero = () => {
 
   const handleCloseModal = () => {
     setIsOpen(false);
+  };
+
+  const getUserEvents = async () => {
+    try {
+      const { year, month, dayNumber } = currentDate;
+      const timestamp = new Date(year, month - 1, dayNumber).getTime();
+      const response = await get(
+        `http://localhost:3001/api/v1/events/${timestamp}`,
+      );
+      console.log(response);
+      setDayFrames(response);
+      setUpdate(false);
+    } catch (error) {
+      console.log(error);
+    }
   };
 
   const customStyles = {
@@ -36,27 +61,28 @@ const Hero = () => {
     },
   };
 
+  useEffect(() => {
+    getUserEvents();
+  }, [currentDate.year, currentDate.month, updateCalendar]);
+
   return (
-    <DateProvider>
-      <div className={styles.heroContainer}>
-        <Calendar setDayData={setDayData} openModal={handleOpenModal} />
-        <Modal
-          isOpen={modalIsOpen}
-          onRequestClose={handleCloseModal}
-          style={customStyles}
-        >
-          <button className={button.close} onClick={handleCloseModal}>
-            X
-          </button>
-          <div className={styles.modalContainer}>
-            {dayData.type === 'scheduler' && <Scheduler />}
-            {dayData.type === 'event' && (
-              <EventPanel eventData={dayData.event} />
-            )}
-          </div>
-        </Modal>
-      </div>
-    </DateProvider>
+    <div className={styles.hero}>
+      <DateSelector currentDate={currentDate} setCurrentDate={setCurrentDate} />
+      <Calendar setDayData={setDayData} openModal={handleOpenModal} />
+      <Modal
+        isOpen={modalIsOpen}
+        onRequestClose={handleCloseModal}
+        style={customStyles}
+      >
+        <button className={button.close} onClick={handleCloseModal}>
+          X
+        </button>
+        <div className={styles.modal}>
+          {dayData.type === 'scheduler' && <Scheduler />}
+          {dayData.type === 'event' && <EventPanel eventData={dayData.event} />}
+        </div>
+      </Modal>
+    </div>
   );
 };
