+++ front-end/src/components/Sidebar/index.js
@@ -1,12 +1,7 @@
 import styles from './Sidebar.module.css';
 
 const Sidebar = () => {
-
-  return(
-    <div>
-      
-    </div>
-  );
-}
+  return <div></div>;
+};
 
 export default Sidebar;
